# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hedge Fund AI Agent Council \u2014 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/discover": {
      "get": {
        "summary": "Discover Get",
        "description": "Plain fan-out without refinement \u2014 kept for the CLI / back-compat.",
        "operationId": "discover_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Research topic, e.g. 'AI memory chip demand'",
              "title": "Query"
            },
            "description": "Research topic, e.g. 'AI memory chip demand'"
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Results"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Discover Post",
        "description": "Refine the query, then fan out. May return clarifying questions (interactive).",
        "operationId": "discover_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DiscoveryResult-Output"
                    },
                    {
                      "$ref": "#/components/schemas/Clarify"
                    }
                  ],
                  "title": "Response Discover Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/discover/clarify": {
      "post": {
        "summary": "Discover Clarify",
        "description": "Resume an interactive run with a clarified query (round cap enforced in refine).",
        "operationId": "discover_clarify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DiscoveryResult-Output"
                    },
                    {
                      "$ref": "#/components/schemas/Clarify"
                    }
                  ],
                  "title": "Response Discover Clarify"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataeng/process": {
      "post": {
        "summary": "Dataeng Process",
        "description": "Run Layer 2 (clean + label + theme + persist), then convene the council.",
        "operationId": "dataeng_process",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryResult-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEngReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "summary": "List Items",
        "description": "Read persisted, cleaned + labelled items (newest first), with optional filters.",
        "operationId": "list_items",
        "parameters": [
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by MACRO or MICRO",
              "title": "Stream"
            },
            "description": "Filter by MACRO or MICRO"
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by a market theme, e.g. 'Solar'",
              "title": "Theme"
            },
            "description": "Filter by a market theme, e.g. 'Solar'"
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by a resolved entity, e.g. '$META'",
              "title": "Ticker"
            },
            "description": "Filter by a resolved entity, e.g. '$META'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CleanedItem"
                  },
                  "title": "Response List Items"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trackers/{tracker}/context": {
      "get": {
        "summary": "Tracker Context Endpoint",
        "description": "Best evidence quote for a tracker (theme) + a real 0..1 relevance score (\u00a77.1).",
        "operationId": "tracker_context_endpoint",
        "parameters": [
          {
            "name": "tracker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tracker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/discover/stream": {
      "get": {
        "summary": "Discover Get Stream",
        "description": "Streaming version of GET /discover (plain fan-out, no refinement).",
        "operationId": "discover_get_stream",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Research topic, e.g. 'AI memory chip demand'",
              "title": "Query"
            },
            "description": "Research topic, e.g. 'AI memory chip demand'"
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Results"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Discover Post Stream",
        "description": "Streaming version of POST /discover (refine, then fan out).",
        "operationId": "discover_post_stream",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/discover/clarify/stream": {
      "post": {
        "summary": "Discover Clarify Stream",
        "description": "Streaming version of POST /discover/clarify.",
        "operationId": "discover_clarify_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataeng/process/stream": {
      "post": {
        "summary": "Dataeng Process Stream",
        "description": "Streaming version of POST /dataeng/process (per-item progress).\n\nRegisters a Job so a client that reloads mid-run can reconnect via\n/dataeng/jobs/{id}/stream. The job id is sent as a leading `job` SSE frame.",
        "operationId": "dataeng_process_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryResult-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataeng/jobs": {
      "get": {
        "summary": "Dataeng Jobs",
        "description": "List tracked discovery jobs (running by default) so a reloaded client can find one.",
        "operationId": "dataeng_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job status",
              "default": "running",
              "title": "Status"
            },
            "description": "Filter by job status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dataeng/jobs/{job_id}/stream": {
      "get": {
        "summary": "Dataeng Job Stream",
        "description": "Reconnect to a job: replay its progress so far, then stream live to completion.",
        "operationId": "dataeng_job_stream",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/items/stream": {
      "get": {
        "summary": "List Items Stream",
        "description": "Streaming version of GET /items (emits a query step, then the rows).",
        "operationId": "list_items_stream",
        "parameters": [
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by MACRO or MICRO",
              "title": "Stream"
            },
            "description": "Filter by MACRO or MICRO"
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by a market theme, e.g. 'Solar'",
              "title": "Theme"
            },
            "description": "Filter by a market theme, e.g. 'Solar'"
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by a resolved entity, e.g. '$META'",
              "title": "Ticker"
            },
            "description": "Filter by a resolved entity, e.g. '$META'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Clarify": {
        "properties": {
          "original_query": {
            "type": "string",
            "title": "Original Query"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          }
        },
        "type": "object",
        "required": [
          "original_query",
          "questions",
          "round"
        ],
        "title": "Clarify",
        "description": "Refinement needs the Commander to clarify (interactive mode, Req 2.5)."
      },
      "ClarifyRequest": {
        "properties": {
          "clarified_query": {
            "type": "string",
            "title": "Clarified Query"
          },
          "round": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Round",
            "default": 1
          },
          "mode": {
            "$ref": "#/components/schemas/ClarificationMode",
            "default": "interactive"
          },
          "max_results": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Results"
          }
        },
        "type": "object",
        "required": [
          "clarified_query"
        ],
        "title": "ClarifyRequest"
      },
...
      "DataEngReport": {
        "properties": {
          "persisted": {
            "type": "integer",
            "title": "Persisted",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/ProcessingFailure"
            },
            "type": "array",
            "title": "Failures"
          }
        },
        "type": "object",
        "title": "DataEngReport",
        "description": "Batch outcome of a Data Engineering run (Req 7.4, 7.5)."
      },
...
      "DiscoverRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "mode": {
            "$ref": "#/components/schemas/ClarificationMode",
            "default": "interactive"
          },
          "max_results": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Results"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoverRequest"
      },
      "DiscoveryResult-Input": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/SkippedSource"
            },
            "type": "array",
            "title": "Skipped"
          },
          "original_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Query"
          },
          "refinement_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refinement Note"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryResult",
        "description": "What the Discovery agent returns for a single topic query."
      },
      "DiscoveryResult-Output": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/SkippedSource"
            },
            "type": "array",
            "title": "Skipped"
          },
          "original_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Query"
          },
          "refinement_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refinement Note"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryResult",
        "description": "What the Discovery agent returns for a single topic query."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}